Implement Fluid Template Engine in index.html and index.php
Replaced the existing Twig templating system with the Fluid templating engine. The Fluid template 'index.html' was created with the inclusion of Bootstrap CSS, and the 'index.php' file was updated accordingly to utilize the Fluid View TemplateView. Removed faker data creation as well as Twig-related code from 'index.php'.